fix(ci): use setuptools to get current package version
Since esp-idf-diag utilizes pyproject.toml instead of setup.py,
retrieving the current package version in the PyPI upload script via
setup.py does not work. This was overlooked in the initial version of
the esp-idf-diag package. Switch to using setuptools directly.
Additionally, use ubuntu-latest and the latest Python version, as
backward compatibility is unnecessary for the PyPI release script. This
approach is inspired by esp-coredump.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>