fix: display correctly built-in recipes information in the list command
The list command provides information on whether a recipe is built-in by
comparing the recipe's resolved path to the BUILTIN_RECIPES_PATH
directory, which contains built-in recipes. If BUILTIN_RECIPES_PATH is
not resolved, built-in recipes might be incorrectly reported as
non-built-in, especially when there are symlinks in the path, such as
lib64->lib. For instance, "idf5.5_py3.12_env/lib64/python3.12" versus
"idf5.5_py3.12_env/lib/python3.12". To correct this, resolve the
BUILTIN_RECIPES_PATH.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>