feat: allow recipes to use the target serial port
To obtain information about the target, the device's serial port must be
specified. A new `--port` option has been added to explicitly define the
port on the command line. If this option is not used, the port is
automatically detected in a way similar to idf.py. The serial port value
is accessible to recipes through a new `PORT` variable. Since this
variable might not be set due to the port's potential unavailability, a
new `port` key has been added to the step definition.  When this key is
set to `true`, the step requires the serial port to be specified;
otherwise, the step is skipped.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>