fix: do not report a failed exec command as an error
Currently, any failed exec command is reported to the user as an error,
which might be confusing since command failure is anticipated. The
purpose of notifying the user about a failed exec command is to indicate
that its output will not be included in the report. Change this
notification to a warning, similar to other commands.

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>